move login endpoint outside of authenticated_router
move login endpoint outside of authenticated_router

move login endpoint outside authenticated router
